+++ feature/work/domain/src/main/java/com/andef/mycarandef/work/domain/usecases/GetWorksByCarIdUseCase.kt
@@ -4,5 +4,5 @@ import com.andef.mycarandef.work.domain.repository.WorkRepository
 import javax.inject.Inject
 
 class GetWorksByCarIdUseCase @Inject constructor(private val repository: WorkRepository) {
-    operator fun invoke(carId: Int) = repository.getWorksByCarId(carId)
+    operator fun invoke(carId: Long) = repository.getWorksByCarId(carId)
 }
